Added more params.
Cleaned up "radius_effects" to not need a global var. We just read the param directly.
We now correctly reset amp or release when the radius effects the other.
Made a param for cutoff. Removed global var and ENC 1 now just sets the param directly. Changing the param will change the engine.