Midi code can now handle polyphonic sequences.
Notes are turned on when a circle bursts but are currently sustained until they burst again. To get around this, you can adjust the envelope on your midi device so that you never enter the sustain section. Will need to add some kind of setting to change the length of the sustain.

Issue #6